requirements.txt
Django
gunicorn
django-heroku
django-bootstrap4
django-allauth
Pillow
django-crispy-forms